pyproject.toml
[project]
name = "whca-graphic"
version = "0.1.0"
description = "Data journalism project: whca-graphic"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "jupyterlab>=4.4.5",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "python-dotenv>=1.0.0",
    "requests>=2.32.4",
    "ipykernel>=6.29.0",
]

[project.optional-dependencies]
viz = [
    "ipywidgets>=8.1.7",
    "matplotlib>=3.10.5",
    "plotly>=6.2.0",
    "seaborn>=0.13.2",
    "altair>=5.2.0",
]
geo = [
    "censusdis>=1.4.0",
    "folium>=0.15.0",
    "geopandas>=1.0.0",
]
data = [
    "polars>=1.0.0",
    "pyarrow>=18.0.0",
    "duckdb>=1.0.0",
]
scraping = [
    "beautifulsoup4>=4.12.0",
    "httpx>=0.27.0",
    "lxml>=5.3.0",
    "selenium>=4.25.0",
    "playwright>=1.49.0",
]
io = [
    "openpyxl>=3.1.0",
    "pyyaml>=6.0.0",
]
databases = [
    "psycopg[binary]>=3.1.0",
    "sqlalchemy>=2.0.0",
]
notebooks = [
    "nbconvert>=7.16.4",
    "notebook>=7.4.5",
]
datawrapper = [
    "datawrapper>=0.9.0",
]
diagnostics = [
    "memory-profiler>=0.61.0",
    "psutil>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["utils"]
